Finished the animation implementation
Added the rest of the code so that each bone in each frame is animated, and added a limitation on how much the bone can displace, so that the animation isn't too weird. On the Blender app, I have to get rid of the links between the bones because when I change a certain bone location, its children and parents also are changed.